Output current state of each timestep into a file an fix some bugs

+++ src/config.h
@@ -5,20 +5,20 @@
 #define CELL_NB 16
 
 typedef struct {
-    // 
-    bool spec_from_file;
-    // space time params
-    int coef_d2;
-    float dt;
-    // demographic params
-    int max_spec;
-    int nspec;
-    int max_fam;
-    int max_pop;
-    int init_fam;
-    int init_fam_size;
-    int init_pop;
-    int ressources;
+  // 
+  bool spec_from_file;
+  // space time params
+  float coef_d2;
+  float dt;
+  // demographic params
+  int max_spec;
+  int nspec;
+  int max_fam;
+  int max_pop;
+  int init_fam;
+  int init_fam_size;
+  int init_pop;
+  int ressources;
 } config;
 
 #endif
